requirements.txt
pyTelegramBotAPI==4.16.1  # (возможно, потребуется более новая версия)
pandas==2.2.1  # или выше (но не факт, что соберётся)
numpy==1.26.4  # (или 2.0.0, если выйдет поддержка)
python-dotenv==1.0.1  # должен работать
schedule==1.2.1  # должен работать